Refactor uiblungs package directory structure
