--- requirements.txt
@@ -1,6 +1,0 @@
-numpy==1.22.4
-opencv-python==4.5.5.64
-pandas==1.4.2
-python-dateutil==2.8.2
-pytz==2022.1
-six==1.16.0
