+++ src/covid_detection/uiblungs/_box/box.py
