+++ src/covid_detection/uiblungs/_separator/separator.py
@@ -1,7 +1,7 @@
 import abc
 import numpy as np
 
-from uiblungs.box.box import ImageBox
+from box.box import ImageBox
 
 
 class LungSeparator(abc.ABC):
