+++ src/covid_detection/uiblungs/_slicer/slicer.py
@@ -1,7 +1,6 @@
 import abc
-import numpy as np
 
-from uiblungs.box.box import ImageBox
+from box.box import ImageBox
 
 
 class BoxSlicer(abc.ABC):
