--- uiblungs/setup.py
@@ -1,8 +1,0 @@
-from setuptools import setup, find_packages
-
-
-setup(
-    name="foo",
-    version="1.0",
-    packages=find_packages(),
-)
